requirements.txt
inquirer~=2.8.0
psutil~=5.8.0
PyYAML~=5.4.1
requests~=2.25.1
pyperclip~=1.8.2
setuptools~=57.0.0